Simplified code generation
Instead of setting up a list of patterns (for funArgs) and then
pattern-matching on them again in our loop, let's just generate a list
of names first and only turn them into patterns when defining funArgs.
This simplifies the code a bit but also prepares some further processing
on the list of names.