Made makeCata actually generate catamorphisms
It turns out that the makeCata function very accurately reflected my
lack of understanding of what constitutes a catamorphism. My thinking
was that a catamorphism is basically a function which replaces all
constructors with a custom function, e.g. for

  data T = X | Y Int | Z T

we have

  expr :: r           -- X constructor
       -> (Int -> r)  -- Y constructor
       -> (T -> r)    -- Z constructor
       -> T
       -> r

However, it turns out that this is *not* a catamorphism: the 'Z'
constructor makes this a recursive data type. A catamorphism would
rather replace *all* calls to the 'T' type constructor, i.e. it would
be

  expr :: r           -- X constructor
       -> (Int -> r)  -- Y constructor
       -> (r -> r)    -- Z constructor
       -> T
       -> r

This complicates 'makeCata' a bit since we have to inspect the
data constructors to see whether a recursive application of the
catamorphism is required. The advantage (besides doing what the module
name suggests...) is that client code no longer needs to recurse
manually.