Introduced some tests
This verifies that the code generated by makeCata compiles and does what
one would expect.

It appears that starting with GHC 7.10, the KindSignatures extension is
required when generating catamorphisms for polymorphic types (the
reasons for this are beyond me at this point, but GHC suggests to use
this extension in its error message so who am I to argue).